backend/pyarud-back/app/config.py
"""
Application Configuration
"""
import os
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()


class Config:
    """Base configuration"""
    
    # Flask Settings
    SECRET_KEY = os.environ.get('SECRET_KEY') or 'dev-secret-key-change-in-production'
    DEBUG = os.environ.get('FLASK_DEBUG', 'False').lower() == 'true'
    
    # CORS Settings
    CORS_ORIGINS = os.environ.get('CORS_ORIGINS', '*').split(',')
    
    # Application Settings
    MAX_CONTENT_LENGTH = 16 * 1024  # 16 KB max request size
    JSON_AS_ASCII = False  # Support for Arabic characters in JSON
    
    # PyArud Settings
    MAX_VERSES_PER_REQUEST = int(os.environ.get('MAX_VERSES_PER_REQUEST', '50'))


class DevelopmentConfig(Config):
    """Development configuration"""
    DEBUG = True


class ProductionConfig(Config):
    """Production configuration"""
    DEBUG = False


# Configuration dictionary
config = {
    'development': DevelopmentConfig,
    'production': ProductionConfig,
    'default': DevelopmentConfig
}
