backend/pyarud-back/requirements.txt
# Core Framework
Flask==3.0.0
Werkzeug==3.0.1

# PyArud - Arabic Poetry Analysis Library
pyarud==0.1.10

# CORS Support for Frontend Communication
Flask-CORS==4.0.0

# Environment Variables Management
python-dotenv==1.0.0

# Input Validation
marshmallow==3.20.1

# Production Server
gunicorn==21.2.0

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
